working code for creating the mysql container.

+++ etc/longshore/dockerfiles/ptp/mysql/docker-entrypoint.sh
@@ -1,48 +1,29 @@
 #!/bin/bash
 set -e
 
-# If the command passed to docker run/exec starts with a - or -- then
-# set the command to mysqld and pass whatever follows as an argument to
-# mysqld.
-if [ "${1:0:1}" = '-' ]; then
-	set -- mysqld "$@"
-fi
+sql_init_file=/root/temp.sql
 
 if [ "$1" = 'mysqld' ]; then
-	# read DATADIR from the MySQL config
-	DATADIR="$("$@" --verbose --help 2>/dev/null | awk '$1 == "datadir" { print $2; exit }')"
-	
-	if [ ! -d "$DATADIR/mysql" ]; then
-		if [ -z "$MYSQL_ROOT_PASSWORD" ]; then
-			echo >&2 'error: database is uninitialized and MYSQL_ROOT_PASSWORD not set'
-			echo >&2 '  Did you forget to add -e MYSQL_ROOT_PASSWORD=... ?'
-			exit 1
-		fi
+  if [ -f "/root/.my.cnf" ]; then
+    # initialize root password.
+    password=$(grep password /root/.my.cnf | cut -d= -f2)
+    if [ -z "$password" ]; then
+      echo "Failed to extract password from /root/my.cnf"
+      exit 1
+    fi
+    cat > "$sql_init_file" <<-EOSQL
+      DELETE FROM mysql.user WHERE user = 'root' ;
+      CREATE USER 'root'@'%' IDENTIFIED BY "$password" ;
+      GRANT ALL ON *.* TO 'root'@'%' WITH GRANT OPTION ;
+      DROP DATABASE IF EXISTS test ;
+EOSQL
+    
+    chown mysql:mysql "$sql_init_file"
 
-	  	
-		echo 'Running mysql_install_db ...'
-		mysql_install_db --datadir="$DATADIR" --basedir=/usr 2>&1 >/dev/null
-		echo 'Finished mysql_install_db'
-		
-		# These statements _must_ be on individual lines, and _must_ end with
-		# semicolons (no line breaks or comments are permitted).
-		# TODO proper SQL escaping on ALL the things D:
-		
-    tempSqlFile=$(mktemp)
-    chown mysql:mysql "$tempSqlFile"
-		cat > "$tempSqlFile" <<-EOSQL
-			DELETE FROM mysql.user ;
-			CREATE USER 'root'@'%' IDENTIFIED BY '${MYSQL_ROOT_PASSWORD}' ;
-			GRANT ALL ON *.* TO 'root'@'%' WITH GRANT OPTION ;
-			DROP DATABASE IF EXISTS test ;
-		EOSQL
-		
-		echo 'FLUSH PRIVILEGES ;' >> "$tempSqlFile"
-		
-		set -- "$@" --init-file="$tempSqlFile"
-	fi
-  
-	chown -R mysql:mysql "$DATADIR"
+    echo 'FLUSH PRIVILEGES ;' >> "$sql_init_file"
+    
+    set -- "$@" --init-file="$sql_init_file"
+  fi
 fi
 
 exec "$@"
